Renamed and reorganized globalConstants
