+++ lib/constants/colors.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
 
-const appTitle = 'Dispatcher';
-
-class FigmaColors {
+class AppColors {
   static const black = Colors.black;
   static const deepDarkBlue = Color.fromRGBO(38, 33, 70, 1.0);
   static const mediumBlue = Color.fromRGBO(90, 90, 137, 1);
@@ -12,14 +10,4 @@ class FigmaColors {
   static const grey = Color.fromRGBO(217, 219, 233, 1);
   static const error = Color.fromRGBO(255, 107, 87, 1);
   static const badgeColor = Color.fromRGBO(217, 219, 233, 1);
-}
-
-class ValidRoutes {
-  static const onboardingScreen = '/onboarding_screen';
-  static const signupLoginScreen = '/sign_up_login_screen';
-}
-
-enum IconDirection {
-  start,
-  end,
 }
