+++ lib/constants/constant_vars.dart
@@ -1,0 +1,1 @@
+const appTitle = 'Dispatcher';
