+++ lib/constants/routes.dart
@@ -1,0 +1,4 @@
+class ValidRoutes {
+  static const onboardingScreen = '/onboarding_screen';
+  static const signupLoginScreen = '/sign_up_login_screen';
+}
