+++ lib/main.dart
@@ -1,11 +1,13 @@
-import 'package:dispatcher/providers/onboarding_step_provider.dart';
 import 'package:flutter/material.dart';
-import 'package:dispatcher/screens/onboarding_screen.dart';
-import 'package:dispatcher/screens/splash_screen.dart';
+
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 
-import './global_constants.dart';
+import './constants/constant_vars.dart';
+import './constants/routes.dart';
+import './providers/onboarding_step_provider.dart';
+import './screens/onboarding_screen.dart';
+import './screens/splash_screen.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
